+++ registry/features/blog/blog-index.ts
@@ -1,0 +1,3 @@
+export * from "./schemas/posts.table";
+export { PostRepository, type Post } from "./post.repository";
+export { BlogService } from "./blog.service";