+++ registry/features/blog/blog-page-server.ts
@@ -1,0 +1,5 @@
+import { BlogService } from "../../features/blog";
+
+export async function load() {
+	return { posts: await BlogService.list() };
+}